Moving pyFROG helper files to separate folder
